moved gad reader to separate file
